Add TimestampedLogFormatter with color support
Refactored colouring functionality of ConsoleLogFormatter into abstract
class.
TimestampedLogFormatter replaces LogFileFormatter with color support.
Marked FileLogFormatter as deprecated.
