+++ src/org/sosy_lab/common/log/PackageSanityTest.java
@@ -13,6 +13,7 @@ import com.google.common.testing.AbstractPackageSanityTests;
 import java.util.logging.Formatter;
 import java.util.logging.Handler;
 import java.util.logging.Level;
+import java.util.logging.LogRecord;
 import java.util.logging.SimpleFormatter;
 import org.sosy_lab.common.configuration.Configuration;
 import org.sosy_lab.common.configuration.InvalidConfigurationException;
@@ -24,6 +25,7 @@ public class PackageSanityTest extends AbstractPackageSanityTests {
     setDefault(Handler.class, new StringBuildingLogHandler());
     setDefault(Level.class, Level.ALL);
     setDefault(Formatter.class, new SimpleFormatter());
+    setDefault(LogRecord.class, new LogRecord(Level.ALL, "test"));
 
     setDefault(Configuration.class, Configuration.defaultConfiguration());
     try {
